requirements.txt
numpy
pandas
scipy
mne
mne-bids
openneuro-py
matplotlib
seaborn
